pyproject.toml
[tool.mypy]
files = ["src/"]

[tool.poetry]
name = "test-python-project"
version = "0.1.0"
description = ""
authors = ["John McRaven <john.mcraven@physicsx.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
altair = {extras = ["all"], version = "^5.3.0"}


[tool.poetry.group.dev.dependencies]
pandas-stubs = "^2.2.2.240603"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
